+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -231,7 +231,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K9EHDPR93KX1NR9Y3J9YENSC"),
+  _id: z.string().default("01K9EZ0AZGQG8HE0E468V8BEHF"),
   display_name: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$inboundSchema,
@@ -274,7 +274,7 @@ export const ListDatasourcesData$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ListDatasourcesData
 > = z.object({
-  id: z.string().default("01K9EHDPR93KX1NR9Y3J9YENSC"),
+  id: z.string().default("01K9EZ0AZGQG8HE0E468V8BEHF"),
   displayName: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$outboundSchema,