+++ packages/orq-rc/src/models/operations/retrievecontact.ts
@@ -112,7 +112,7 @@ export const RetrieveContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-11-07T06:52:38.992Z",
+    "2025-11-07T10:50:00.302Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -149,7 +149,7 @@ export const RetrieveContactResponseBody$outboundSchema: z.ZodType<
   tags: z.array(z.string()).optional(),
   metadata: z.record(z.any()).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-11-07T06:52:38.992Z"))
+  updated: z.date().default(() => new Date("2025-11-07T10:50:00.302Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {