+++ packages/orq-rc/RELEASES.md
@@ -4839,3 +4839,13 @@ Based on:
 - [typescript v4.0.0-rc.14] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.14] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.14 - packages/orq-rc
+
+## 2025-11-07 10:51:54
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.650.0 (2.743.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.15] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.15] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.15 - packages/orq-rc